fix(member): #163 반환타입 변경 및 주석 제거

+++ identity-service/src/main/java/shop/dear/identity/member/domain/repository/ArchivedAccountRepository.java
@@ -8,5 +8,5 @@ import java.util.List;
 public interface ArchivedAccountRepository {
 
     void saveAll(List<ArchivedAccount> archivedAccounts);
-    long deleteExpired(LocalDateTime expiresAt);
+    int deleteExpired(LocalDateTime expiresAt);
 }
